feat: add config inference to evaluateSuite in worktree
