+++ harness/evaluate.ts
@@ -7,7 +7,7 @@ import { calculateMetrics } from './lib/metrics.ts';
 import { generateMarkdownReport, generateJsonReport, saveReports } from './lib/reporting.ts';
 import { resultsDir } from '../lib/paths.ts';
 
-import type { SuiteConfig } from './config.ts';
+import { Serving, type SuiteConfig } from './config.ts';
 
 export async function evaluateSuite(suiteResultsDir: string, suiteName: string) {
   console.log(`Evaluating suite: ${suiteName}`.cyan);
@@ -29,7 +29,31 @@ export async function evaluateSuite(suiteResultsDir: string, suiteName: string) 
   }
 
   if (!suiteConfig) {
-    console.error(`⚠️ No suite_config.json found or failed to parse in ${suiteResultsDir}. Aborting evaluation.`.red);
+    console.warn(`⚠️ No suite_config.json found in ${suiteResultsDir}. Inferring config...`.yellow);
+    
+    let agent = 'gemini-cli';
+    let serving: Serving = 'mcp';
+
+    const evalsPath = path.join(suiteResultsDir, 'evals.json');
+    if (fs.existsSync(evalsPath)) {
+      try {
+        const oldEvals = JSON.parse(fs.readFileSync(evalsPath, 'utf8'));
+        if (oldEvals.agent) agent = oldEvals.agent;
+        if (oldEvals.serving) serving = oldEvals.serving;
+        else if (oldEvals.enableSkills !== undefined) {
+          serving = oldEvals.enableSkills ? 'skills' : 'mcp';
+        }
+      } catch {
+        // Ignore parse error
+      }
+    }
+
+    suiteConfig = { agent, serving, tasks: [], name: null, numRuns: 1, mcpServersToEnable: [] };
+    console.log(`Inferred: agent=${agent}, serving=${serving}`.cyan);
+  }
+
+  if (!suiteConfig) {
+    console.error(`⚠️ Failed to infer suite config for ${suiteResultsDir}. Aborting evaluation.`.red);
     return;
   }
